pyproject.toml
[tool.poetry]
name = "cs-demand-model"
version = "0.2.0"
description = "Python port of Social Finance's children's social care demand model"
authors = ["dezog", "tab1tha", "kws"]
packages = [
    { include = "cs_demand_model" },
    { include = "cs_demand_model_samples" },
]

[tool.poetry.dependencies]
numpy = "1.22.4"
pandas = "1.4.2"
python = "^3.10"
python-dateutil = "^2.8.2"
fs = "^2.4.16"
click = "^8.1.3"
PyYAML = "^6.0"
matplotlib = { version = "^3.5.2", optional = true }
tqdm = { version = "^4.64.1", optional = true }
Flask = { version = "^2.1.2", optional = true }
Flask-Cors = { version = "^3.0.10", optional = true }
jupyterlab = { version = "^3.4.8", optional = true }
openpyxl = {version = "^3.0.10", optional = true}
plotly = {version = "^5.11.0", optional = true}
rpc-wrap = {extras = ["cli"], version = "^0.3.0"}


[tool.poetry.dev-dependencies]
pytest = "^7.1"
coverage = "^6.3"
black = "^22.8.0"
isort = "^5.10.1"

[tool.poetry.extras]
cli = ["matplotlib", "tqdm", "openpyxl"]
web = ["Flask", "Flask-Cors"]
jupyter = ["jupyterlab", "matplotlib", "tqdm", "openpyxl", "plotly"]
pyodide = ["openpyxl", "plotly"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
